{{cookiecutter.repo_name}}/requirements/development.txt
# Development requirements

-r base.txt

django-debug-toolbar==2.2
pytest-django==3.9.0
mixer==6.1.3
pytest-cov==2.10.0